pyproject.toml
[tool.poetry]
name = "bavard-ml-utils"
version = "0.2.6"
description = "Utilities for machine learning, python web services, and cloud infrastructure"
license = "MIT"
authors = ["Bavard AI, Inc. <dev@bavard.ai>"]
readme = "README.md"
repository = "https://github.com/bavard-ai/bavard-ml-utils"
documentation = "https://docs-bavard-ml-utils.web.app/"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = ">= 0.63.0, < 1.0.0"
pydantic = "^1.7.3"
numpy = {version = "^1.19.2", optional = true}
scikit-learn = {version = ">= 0.24.2, < 2.0.0", optional = true}
networkx = {version = "^2.6.3", optional = true}
requests = "^2.21.0"
loguru = ">= 0.5.1, < 1.0.0"
google-cloud-storage = {version = "^1.35.1", optional = true}
google-cloud-pubsub = {version = "^2.2.0", optional = true}
google-cloud-error-reporting = {version = "^1.1.1", optional = true}
google-cloud-firestore = {version = "^2.3.2", optional = true}

[tool.poetry.dev-dependencies]
torch = "^1.9.1"
transformers = "^4.10.2"
tensorflow = "^2.6.0"
sentencepiece = "^0.1.96"
Sphinx = "^4.2.0"
myst-parser = "^0.15.2"
sphinx-rtd-theme = "^1.0.0"
toml = "^0.10.2"
pre-commit = "^2.15.0"
mypy = "^0.910"

[tool.poetry.extras]
ml = ["numpy", "scikit-learn", "networkx"]
gcp = ["google-cloud-storage", "google-cloud-pubsub", "google-cloud-error-reporting", "google-cloud-firestore"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
known_first_party = "bavard_ml_utils,test"

[tool.mypy]
ignore_missing_imports = true
